ref: refactored code and misc updates
