--- lib/lang/index.ts
@@ -1,1 +1,0 @@
-export {}
