--- lib/utils/indian.js
@@ -1,1 +1,0 @@
-export {}
