--- lib/utils/international.js
@@ -1,1 +1,0 @@
-export {}
