+++ src/lang/en.ts
@@ -1,0 +1,9 @@
+export const singleDigit: Array<string> = [
+  '', 'one', 'two', 'three', 'four', 'five', 'six', 'seven', 'eight', 'nine', ''
+]
+export const twoDigit: Array<string> = [
+  '', 'ten', 'twenty', 'thirty', 'fourty', 'fifty', 'sixty', 'seventy', 'eighty', 'ninety', ''
+]
+export const caseOne: Array<string> = [
+  '', 'eleven', 'twelve', 'thirteen', 'fourteen', 'fifteen', 'sixteen', 'seventeen', 'eighteen', 'nineteen', ''
+]
