+++ src/utils/intl.ts
@@ -1,0 +1,15 @@
+export function handleHundredThousands(key: string): string {
+	return ``
+}
+export function handleMillions(key: string): string {
+	return ``
+}
+export function handleTenMillions(key: string): string {
+	return ``
+}
+export function handleBillions(key: string): string {
+	return ``
+}
+export function handleHundredBillions(key: string): string {
+	return ``
+}
